make minor bug fixes to GRangesViewer GUI
I made a few minor bug fixes that I recently discovered, so the GUI
should be a tad more functional. I also changed the error message
to placeholder text (so it disappears when the user sets focus on
the search bar), which is only supported in Qt version 4.7+, so this
feature may need to be removed if we want backwards compatibility.