add slider to RangeParWidget; make ParControlPanel auto-update
There is now a slider in the RangeParWidget that is sync'd with
the spinbox; this should make it easier to change the numeric
variables, without having to type in values.

The ParControlPanel now automatically updates the GraphicPars object
that is passed to it. The submit button is no longer necessary
(commented that out of the code).