update ControlPanel to handle using QColor class
The only noticeable change in the GUI due to changing from the
previous color class that Tengfei was going to implement to QColor
is that the text describing the default colors will be in the
hex-tag form (#000000) as opposed to the color name. Users can
still specify colors with names, though (i.e. 'white', 'green').