add widgets for ColorEnum and GlyphEnum pars to ControlPanel
There are now widgets that create drop-down menus with icons for
GraphicPars of type ColorEnum and GlyphEnum in ControlPanel. There's
currently a bug in ControlPanel that creates extra drop-down menus
for parameters of these types... I just have to think about how to
restructure the class-checking during widget-generation (which needed
to happen at some point anyway).