check in the rest of my code
ModeControlPanel should work fairly well, although many of the
features from ControlPanel won't work because they're not implemented
in this parameter paradigm yet. This file will have to be updated when
things like resetting to defaults, parameter text, and parameter
tooltips become available.

AppWindow-GUI.R contains a start on creating an application window
for visnab. Particularly, at the bottom of the file, there is an
example of how to create a menu for a Group object... was also trying
to show how you could create and change the control panel for the
parameters, but didn't get to the changing part.

SearchBarList-GUI.R is a start on a search bar for a Genomic Ranges
List.